Set up newtypes and organize predef kind of.
