+++ src/main/scala/com/jacoby6000/cloneherodb/predef/package.scala
@@ -1,0 +1,21 @@
+package com.jacoby6000.cloneherodb
+
+
+package object predef
+  extends JavaCompatCancer
+    with ScalaPrimitives
+    with ScalazTypeclassInstances
+    with Syntax
+    with CatsCompat {
+
+  type StringContext = scala.StringContext
+  val StringContext: scala.StringContext.type = scala.StringContext
+
+  def implicitly[A](implicit a: A): A = a
+  def identity[A](a: A): A = a
+}
+
+
+
+
+
